backend/flask_services/parser_service/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
PyMuPDF==1.23.8
python-docx==1.1.0
spacy==3.7.2
gunicorn==21.2.0
werkzeug==3.0.1

